# Try async render