Before

```
This is ApacheBench, Version 2.3 <$Revision: 1706008 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking annon-status-web.herokuapp.com (be patient)


Server Software:        Cowboy
Server Hostname:        annon-status-web.herokuapp.com
Server Port:            443
SSL/TLS Protocol:       TLSv1.2,ECDHE-RSA-AES128-GCM-SHA256,2048,128

Document Path:          /
Document Length:        57995 bytes

Concurrency Level:      100
Time taken for tests:   23.037 seconds
Complete requests:      1000
Failed requests:        880
   (Connect: 0, Receive: 0, Length: 880, Exceptions: 0)
Total transferred:      16462680 bytes
HTML transferred:       16094680 bytes
Requests per second:    43.41 [#/sec] (mean)
Time per request:       2303.687 [ms] (mean)
Time per request:       23.037 [ms] (mean, across all concurrent requests)
Transfer rate:          697.87 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:      189  681 526.8    476    4717
Processing:   237 1490 2135.6    672   11427
Waiting:      233 1194 1416.6    600    9734
Total:        466 2171 2265.0   1577   12464

Percentage of the requests served within a certain time (ms)
  50%   1577
  66%   1909
  75%   2219
  80%   2592
  90%   3864
  95%   8411
  98%  11002
  99%  11600
 100%  12464 (longest request)
```

After

```
This is ApacheBench, Version 2.3 <$Revision: 1706008 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking annon-status-web-6.herokuapp.com (be patient)


Server Software:        Cowboy
Server Hostname:        annon-status-web-6.herokuapp.com
Server Port:            443
SSL/TLS Protocol:       TLSv1.2,ECDHE-RSA-AES128-GCM-SHA256,2048,128

Document Path:          /
Document Length:        494 bytes

Concurrency Level:      100
Time taken for tests:   11.939 seconds
Complete requests:      1000
Failed requests:        0
Non-2xx responses:      1000
Total transferred:      669000 bytes
HTML transferred:       494000 bytes
Requests per second:    83.76 [#/sec] (mean)
Time per request:       1193.931 [ms] (mean)
Time per request:       11.939 [ms] (mean, across all concurrent requests)
Transfer rate:          54.72 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:      441  862 203.7    847    2009
Processing:   120  253 129.3    205     806
Waiting:      120  217  92.9    185     648
Total:        568 1115 193.4   1072    2207

Percentage of the requests served within a certain time (ms)
  50%   1072
  66%   1091
  75%   1143
  80%   1182
  90%   1415
  95%   1562
  98%   1631
  99%   1713
 100%   2207 (longest request)
```